% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-formulas.R
\name{bayes_R2_res}
\alias{bayes_R2_res}
\title{Bayes R2 Res}
\usage{
bayes_R2_res(y, ypred)
}
\arguments{
\item{y}{(numeric) y variable}

\item{ypred}{(numeric) predicted y from fitted formulas object}
}
\description{
Bayes R2 Res
}
