% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelDiagnostics.R
\name{convergenceDiagnostics}
\alias{convergenceDiagnostics}
\title{Convergence Diagnostics}
\usage{
convergenceDiagnostics(parameters, nChains)
}
\arguments{
\item{parameters}{parameters}

\item{nChains}{number of chains}
}
\description{
Convergence Diagnostics
}
