% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-formulas.R
\name{enrichForm}
\alias{enrichForm}
\title{Enrich Form}
\usage{
enrichForm(formDF, parNames, formulasObject, y)
}
\arguments{
\item{formDF}{(data.frame) a dataframe containing the form, name, x and y variables}

\item{parNames}{(character) names of parameters from form}

\item{formulasObject}{(list) fitted formulas object}

\item{y}{(numeric) y variable}
}
\description{
Enrich Form
}
