% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensities.R
\name{extractQuantileProbs}
\alias{extractQuantileProbs}
\title{Extract Quantile Props}
\usage{
extractQuantileProbs(prop = 0.8)
}
\arguments{
\item{prop}{(numeric) double between 0 and 1: length of credibility interval. The default value is 80 percent.}
}
\description{
Extract lower and upper probabilities for the \code{stats::quantile} function
}
