% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-module-remoteModels.R
\name{getRemoteModelsFromGithub}
\alias{getRemoteModelsFromGithub}
\title{Get Remote Models From Github}
\usage{
getRemoteModelsFromGithub(githubRepo, apiOut = NULL)
}
\arguments{
\item{githubRepo}{(character) name of used github repository}

\item{apiOut}{(list) output of api call from getGithubContent}
}
\description{
Get remote models from github directory
}
