% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensities.R
\name{setDefaultTitle}
\alias{setDefaultTitle}
\title{Set Deafult Title}
\usage{
setDefaultTitle(g, type, plotType)
}
\arguments{
\item{g}{(ggplot) ggplot}

\item{type}{A character. Show distribution by "Sample" (default), "Category" (only if supplied by yEstimates) or all individuals "Combined"?}

\item{plotType}{A character. Form of presentation, choices are "KernelDensity" (default), "Histogram" or "Boxplot"}
}
\description{
Set Deafult Title
}
