% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startApplication.R
\name{startApplication}
\alias{startApplication}
\title{Start Application}
\usage{
startApplication(
  port = 4242,
  launch.browser = getOption("shiny.launch.browser", interactive())
)
}
\arguments{
\item{port}{port of web application}

\item{launch.browser}{whether to launch the browser}
}
\description{
Start Application
}
