% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensities.R
\name{summariseEstimates}
\alias{summariseEstimates}
\title{Summarise estimates and check for pairwise differences}
\usage{
summariseEstimates(
  yEstimates,
  type = "Individual",
  checkDifferences = TRUE,
  probability = 0.95,
  checkDifferencesReference = FALSE,
  referenceType = "dist",
  referenceDist = "norm",
  referenceParameters = c(0, 1),
  referenceSample = c(),
  referenceTable = matrix(data = NA, nrow = 2),
  meanType = "1"
)
}
\arguments{
\item{yEstimates}{An object created by \code{\link{estimateY}}. Distributions of the dependent variable.}

\item{type}{A character. Summarise and compare distribution by "Individual" (default), "Category" or all individuals "Combined"?}

\item{checkDifferences}{A boolean. Check pairwise differences TRUE / FALSE. Only relevant if type = "Individual" or "Category"}

\item{probability}{A numeric. Level of the credible intervals.}

\item{checkDifferencesReference}{A boolean. Check differences with reference distribution or sample.}

\item{referenceType}{A character. Type of reference, one of: "dist", "sample" and "freqTable"}

\item{referenceDist}{A character. Type of distribution in R, defaults to "norm" (if referenceType = "dist")}

\item{referenceParameters}{A numeric vector. Parameters of reference distribution (same order as in corresponding r-function)}

\item{referenceSample}{A numeric vector. Sample from reference distribution (if referenceType = "sample")}

\item{referenceTable}{A numeric vector. Sample from reference distribution (if referenceType = "sample")}

\item{meanType}{A character "1" for total variance for category type, "2" for mean variance}
}
\value{
A ggplot2 object with densities
}
\description{
Summarise estimates and check for pairwise differences
}
