% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-module-downAndUploadModel.R
\name{uploadModelUI}
\alias{uploadModelUI}
\alias{uploadModel}
\title{Upload model module}
\usage{
uploadModelUI(id, label)

uploadModel(input, output, session)
}
\arguments{
\item{id}{id of module}

\item{label}{label of module}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\description{
UI function to upload a zip file with notes and a list of models

Backend for upload model module
}
