% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-module-remoteModels.R
\name{getRemoteModelsFromGithub}
\alias{getRemoteModelsFromGithub}
\title{Get Remote Models From Github}
\usage{
getRemoteModelsFromGithub(githubRepo)
}
\arguments{
\item{githubRepo}{(character) name of used github repository}
}
\description{
Get remote models from github directory
}
