% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensities.R
\name{plotDensities}
\alias{plotDensities}
\title{Plot the densities estimated by \code{\link{estimateY}}.}
\usage{
plotDensities(
  yEstimates,
  type = "Sample",
  plotType = "KernelDensity",
  nBins = 15,
  meanType = "1",
  ylabel = "",
  xlabel = "",
  headerLabel = "",
  xTextSize = 24,
  yTextSize = 24,
  xAxisSize = 18,
  yAxisSize = 18,
  showLegend = TRUE,
  whiskerMultiplier = 0.95,
  boxQuantile = 0.68
)
}
\arguments{
\item{yEstimates}{An object created by \code{\link{estimateY}}. Distributions of the dependent variable.}

\item{type}{A character. Show distribution by "Sample" (default), "Category" (only if supplied by yEstimates) or all individuals "Combined"?}

\item{plotType}{A character. Form of presentation, choices are "KernelDensity" (default), "Histogram" or "Boxplot"}

\item{nBins}{An integer Number of bins for histogram}

\item{meanType}{A character "1" for total variance for category type, "2" for mean variance}

\item{ylabel}{ylabel}

\item{xlabel}{xlabel}

\item{headerLabel}{title label}

\item{xTextSize}{x label text size}

\item{yTextSize}{y label text size}

\item{xAxisSize}{x axis text size}

\item{yAxisSize}{y axis text size}

\item{showLegend}{show legend?}

\item{whiskerMultiplier}{multiplier of boxplot whiskers}

\item{boxQuantile}{quantile of boxplot}
}
\value{
A ggplot2 object with densities
}
\description{
Plot the densities estimated by \code{\link{estimateY}}.
}
