% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensities.R
\name{plotFunctions}
\alias{plotFunctions}
\title{Plot functions}
\usage{
plotFunctions(data, xVar, yVar, object, PointSize = 1, LineWidth = 1)
}
\arguments{
\item{data}{data}

\item{xVar}{independent variable}

\item{yVar}{dependent variable}

\item{object}{object}

\item{PointSize}{point size}

\item{LineWidth}{line width}
}
\description{
Plot functions
}
