% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensities.R
\name{plotFunctions}
\alias{plotFunctions}
\title{Plot functions}
\usage{
plotFunctions(data, xVar, yVar, object)
}
\arguments{
\item{data}{data}

\item{xVar}{independent variable}

\item{yVar}{dependent variable}

\item{object}{object}
}
\description{
Plot functions
}
