% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensities.R
\name{plotFunctions}
\alias{plotFunctions}
\title{Plot functions}
\usage{
plotFunctions(
  data,
  xVar,
  yVar,
  object,
  PointSize = 1,
  LineWidth = 1,
  prop = 0.8,
  ...
)
}
\arguments{
\item{data}{data}

\item{xVar}{independent variable}

\item{yVar}{dependent variable}

\item{object}{object}

\item{PointSize}{point size}

\item{LineWidth}{line width}

\item{prop}{(numeric) double between 0 and 1: length of credibility interval. The default value is 80 percent.}

\item{...}{Other arguments passed on to \code{layer()} for display of the credibility interval.
These are often aesthetics, used to set an aesthetic to a fixed value, like 
\code{colour = "red"} or \code{alpha = 0.5}.}
}
\description{
Plot functions
}
