% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-module-exportPopUp.R, R/NAMESPACE.R,
%   R/modelDiagnostics.R
\name{exportPlotPopUpUI}
\alias{exportPlotPopUpUI}
\alias{exportPlotPopUpServer}
\alias{shinyModule}
\alias{textExportButton}
\alias{textExport}
\title{Server and UI Functions for Shiny Module}
\usage{
exportPlotPopUpUI(id)

exportPlotPopUpServer(id, exportPlot, filename)

textExportButton(id, title = "Download")

textExport(input, output, session, printFun, filename = "output")
}
\arguments{
\item{id}{namespace id}

\item{exportPlot}{plot to be exported}

\item{filename}{(character) file name}

\item{title}{title of tab in tabset panel}

\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session}

\item{printFun}{print function}
}
\description{
Server and UI Functions for Shiny Module

Text Export Button UI

Text Export Server
}
