% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMESPACE.R, R/modelDiagnostics.R
\name{shinyModule}
\alias{shinyModule}
\alias{textExportButton}
\alias{textExport}
\title{Server and UI Functions for Shiny Module}
\usage{
textExportButton(id, title = "Download")

textExport(input, output, session, printFun, filename = "output")
}
\arguments{
\item{id}{namespace id}

\item{title}{title of tab in tabset panel}

\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session}

\item{printFun}{print function}

\item{filename}{(character) file name}
}
\description{
Server and UI Functions for Shiny Module

Text Export Button UI

Text Export Server
}
