% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-module-downAndUploadModel.R
\name{uploadModelUI}
\alias{uploadModelUI}
\alias{uploadModel}
\title{Upload model module}
\usage{
uploadModelUI(id, label)

uploadModel(input, output, session, yEstimates, formulas, data, uploadedNotes)
}
\arguments{
\item{id}{id of module}

\item{label}{label of module}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{yEstimates}{(reactive) An object created by \code{\link{estimateY}}.
Distributions of the dependent variable.}

\item{formulas}{(reactive) formulas}

\item{data}{(reactive) data}

\item{uploadedNotes}{(reactive) variable that stores content of README.txt}
}
\description{
UI function to upload a zip file with notes and a list of models

Backend for upload model module
}
